move short number utils to shared
