+++ packages/shared/src/index.ts
@@ -30,4 +30,3 @@ export * from "./utils/constants";
 export * from "./utils/types";
 export * from "./utils/opaque-types";
 export * from "./utils/enums";
-export * from "./utils/formatters";
